close side panel when selected dataset is deleted
